Plot the binary tree built using CytomeTree.

Plot the binary tree built using CytomeTree.

Usage

plot_graph(CytomeTreeObj, Ecex = 1, Ecolor = 8, Vcex = 0.8, Vcolor = 0, ...)

Arguments

CytomeTreeObj

An object of class CytomeTree.

Ecex

Number indicating the amount by which text on the edges should be scaled. Default is 1.

Ecolor

An integer or a string of character to color edges of the graph. Default is 8.

Vcex

Number indicating the amount by which text in the vertices should be scaled. Default is .8.

Vcolor

A vector of class numeric or character to color vertices of the graph. Default is 0.

...

additional arguments to be passed to plot_graph

Author

Chariff Alkhassim